RHINENG-8133: rename and unify structures for /systems
