+++ manager/controllers/advisory_systems.go
@@ -12,8 +12,51 @@ import (
 	"gorm.io/gorm"
 )
 
+var AdvisorySystemsFields = database.MustGetQueryAttrs(&AdvisorySystemDBLookup{})
+var AdvisorySystemsSelect = database.MustGetSelect(&AdvisorySystemDBLookup{})
+
+type AdvisorySystemsResponse struct {
+	Data  []AdvisorySystemItem `json:"data"`
+	Links Links                `json:"links"`
+	Meta  ListMeta             `json:"meta"`
+}
+
+type AdvisorySystemItem struct {
+	Attributes AdvisorySystemItemAttributes `json:"attributes"`
+	ID         string                       `json:"id"`
+	Type       string                       `json:"type"`
+}
+
+// nolint: lll
+type AdvisorySystemDBLookup struct {
+	SystemsMetaTagTotal
+	AdvisorySystemItemAttributes
+	SystemIDAttribute
+}
+
+// nolint: lll
+type AdvisorySystemItemAttributes struct {
+	SystemDisplayName
+	SystemLastUpload
+	SystemStale
+	OSAttributes
+	SystemTimestamps
+	SystemTags
+	SystemGroups
+	BaselineIDAttr
+	BaselineNameAttr
+	SystemAdvisoryStatus
+	SystemSatelliteManaged
+	SystemBuiltPkgcache
+}
+
+type SystemsStatusID struct {
+	SystemsID
+	SystemAdvisoryStatus
+}
+
 var AdvisorySystemOpts = ListOpts{
-	Fields: SystemsFields,
+	Fields: AdvisorySystemsFields,
 	// By default, we show only fresh systems. If all systems are required, you must pass in:true,false filter into the api
 	DefaultFilters: map[string]FilterData{
 		"stale": {
@@ -51,7 +94,7 @@ func advisorySystemsCommon(c *gin.Context) (*gorm.DB, *ListMeta, []string, error
 	}
 
 	query := buildAdvisorySystemsQuery(db, account, groups, advisoryName)
-	opts := AdvisorySystemOptsV3
+	opts := AdvisorySystemOpts
 	filters, err := ParseAllFilters(c, opts)
 	if err != nil {
 		return nil, nil, nil, err
@@ -90,7 +133,7 @@ func advisorySystemsCommon(c *gin.Context) (*gorm.DB, *ListMeta, []string, error
 // @Param    filter[system_profile][ansible][controller_version]	query string 	false "Filter systems by ansible version"
 // @Param    filter[system_profile][mssql]							query string 	false "Filter systems by mssql version"
 // @Param    filter[system_profile][mssql][version]					query string 	false "Filter systems by mssql version"
-// @Success 200 {object} AdvisorySystemsResponseV3
+// @Success 200 {object} AdvisorySystemsResponse
 // @Failure 400 {object} utils.ErrorResponse
 // @Failure 404 {object} utils.ErrorResponse
 // @Failure 500 {object} utils.ErrorResponse
@@ -114,12 +157,60 @@ func AdvisorySystemsListHandler(c *gin.Context) {
 	if err != nil {
 		return // Error handled in method itself
 	}
-	var resp = AdvisorySystemsResponseV3{
+	var resp = AdvisorySystemsResponse{
 		Data:  data,
 		Links: *links,
 		Meta:  *meta,
 	}
 	c.JSON(http.StatusOK, &resp)
+}
+
+func buildAdvisorySystemsData(fields []AdvisorySystemDBLookup) ([]AdvisorySystemItem, int) {
+	var total int
+	var err error
+	if len(fields) > 0 {
+		total = fields[0].Total
+	}
+	data := make([]AdvisorySystemItem, len(fields))
+	for i := range fields {
+		if err = parseSystemItems(fields[i].TagsStr, &fields[i].AdvisorySystemItemAttributes.Tags); err != nil {
+			utils.LogDebug("err", err.Error(), "inventory_id", fields[i].ID, "system tags parsing failed")
+		}
+		if err = parseSystemItems(fields[i].GroupsStr, &fields[i].AdvisorySystemItemAttributes.Groups); err != nil {
+			utils.LogDebug("err", err.Error(), "inventory_id", fields[i].ID, "system groups parsing failed")
+		}
+		data[i] = AdvisorySystemItem{
+			Attributes: fields[i].AdvisorySystemItemAttributes,
+			ID:         fields[i].ID,
+			Type:       "system",
+		}
+	}
+	return data, total
+}
+
+func systemsIDsStatus(c *gin.Context, systems []SystemsStatusID, meta *ListMeta) (IDsStatusResponse, error) {
+	var total int
+	resp := IDsStatusResponse{}
+	if len(systems) > 0 {
+		total = systems[0].Total
+	}
+	if meta.Offset > total {
+		err := errors.New("Offset")
+		LogAndRespBadRequest(c, err, InvalidOffsetMsg)
+		return resp, err
+	}
+	if systems == nil {
+		return resp, nil
+	}
+	ids := make([]string, len(systems))
+	data := make([]IDStatus, len(systems))
+	for i, x := range systems {
+		ids[i] = x.ID
+		data[i] = IDStatus{x.ID, x.Status}
+	}
+	resp.IDs = ids
+	resp.Data = data
+	return resp, nil
 }
 
 // nolint: lll
