+++ manager/controllers/systems_auth_test.go
@@ -10,7 +10,7 @@ import (
 
 func testAccountSystemCounts(t *testing.T, acc int, count int) {
 	core.SetupTest(t)
-	var output SystemsResponseV3
+	var output SystemsResponse
 	w := CreateRequestRouterWithAccount("GET", "/", "", "", nil, "", SystemsListHandler, acc)
 	CheckResponse(t, w, http.StatusOK, &output)
 	// data
